requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.0.5
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.8
mysqlclient==1.3.13
pytz==2018.6
requests==2.21.0
urllib3==1.24.1
uWSGI==2.0.18
